Move the [tool.nab] config ladder out of nab-project into nab.config
