+++ nab-project/benchmarks/benchmark_config.py
@@ -7,12 +7,12 @@ from typing import TYPE_CHECKING, NamedTuple
 
 from nab_index.local_index import is_file_url
 from nab_index.multi_index import IndexConfig
-from nab_project.config import PackageOverride
 from nab_project.fetch import DEFAULT_INDEX_NAME, DEFAULT_INDEX_URL, IndexRoute
 from nab_project.inputs import ResolveInputs
 from nab_provider._vendor.packaging.ranges import VersionRange
 from nab_provider._vendor.packaging.requirements import Requirement
 from nab_provider._vendor.packaging.utils import InvalidName, canonicalize_name
+from nab_provider.overrides import PackageOverride
 from nab_provider.provider import (
     BuildPolicy,
     DistPolicy,
