+++ nab-project/tests/property_python/strategies.py
@@ -19,26 +19,16 @@ tests exercise:
 from __future__ import annotations
 
 import os
-from datetime import datetime, timezone
 
 from hypothesis import HealthCheck, settings
 from hypothesis import strategies as st
 
-from nab_project.config import PackageOverride
-from nab_provider._vendor.packaging.requirements import Requirement
 from nab_provider._vendor.packaging.specifiers import SpecifierSet
-from nab_provider._vendor.packaging.utils import canonicalize_name
 from nab_provider._vendor.packaging.version import Version
-from nab_provider.provider import BuildPolicy, DistPolicy
 from nab_provider.tags import PlatformSpec
 from nab_provider.target import ResolveTarget
 
 PACKAGE_NAMES = [f"pkg{i}" for i in range(10)]
-
-# Specifier clauses over a small version grid so two random requirements
-# overlap roughly half the time; a bare name (full range) is included so
-# the always-overlaps case is exercised.
-OVERRIDE_SPECIFIERS = ["", "<=2", ">=2", "<3", ">=3", ">1,<4", "==2"]
 
 VERSION_POOL = [
     Version(f"{major}.{minor}") for major in range(1, 6) for minor in range(3)
@@ -223,43 +213,6 @@ def small_packaging_graphs(
     graph["root"] = {Version("1.0"): root_deps}
 
     return graph
-
-
-@st.composite
-def package_overrides(draw: st.DrawFn, *, name: str) -> PackageOverride:
-    """Draw one per-package ``PackageOverride`` for ``name``.
-
-    The requirement is ``name`` plus a specifier drawn from a small grid
-    so two draws overlap roughly half the time.  The body sets exactly one
-    field, drawn across the policy surfaces, so the overlap property
-    exercises per-field discrimination and the uploaded-prior-to
-    cutoff/disable bucketing (a datetime cutoff and a ``false`` disable
-    count as the same field).
-    """
-    specifier = draw(st.sampled_from(OVERRIDE_SPECIFIERS))
-    requirement = Requirement(f"{name} {specifier}".strip())
-    field = draw(st.sampled_from(["dist", "build", "upload_cutoff", "upload_off"]))
-    dist_policy: DistPolicy | None = None
-    build_policy: BuildPolicy | None = None
-    uploaded_prior_to: datetime | None = None
-    uploaded_prior_to_disabled = False
-    if field == "dist":
-        dist_policy = draw(st.sampled_from(list(DistPolicy)))
-    elif field == "build":
-        build_policy = draw(st.sampled_from(list(BuildPolicy)))
-    elif field == "upload_cutoff":
-        uploaded_prior_to = datetime(2024, 1, 1, tzinfo=timezone.utc)
-    else:
-        uploaded_prior_to_disabled = True
-    return PackageOverride(
-        requirement=requirement,
-        name=canonicalize_name(name),
-        version_range=requirement.specifier.to_range(),
-        dist_policy=dist_policy,
-        build_policy=build_policy,
-        uploaded_prior_to=uploaded_prior_to,
-        uploaded_prior_to_disabled=uploaded_prior_to_disabled,
-    )
 
 
 # A version pool wide enough that random bounds over it land on every
