+++ src/nab/_lock.py
@@ -28,7 +28,6 @@ from typing import TYPE_CHECKING, Any, NamedTuple, NoReturn
 from nab._version import __version__
 from nab.seam import resolve_inputs, resolve_targets
 from nab_project import toml_io
-from nab_project.config import ConfigError, NabProjectConfig, ResolveMode
 from nab_project.lockfile import (
     DisjointnessError,
     DivergentBaseDependencyError,
@@ -89,6 +88,7 @@ from ._run import (
     project_override_arguments,
     read_config_ladder,
 )
+from .config.model import ConfigError, NabProjectConfig, ResolveMode
 from .flagtypes import (  # noqa: TC001 - get_type_hints resolves these at runtime
     BuildPolicyFlag,
     DecisionOrderFlag,
